requirements.txt
streamlit>=1.25
pillow
numpy
tensorflow
scikit-learn
